chore: fixes loggers for tinygo.
