+++ loggers/noop_formater.go
@@ -1,0 +1,5 @@
+package loggers
+
+func noopFormater(al *AuditLog) ([]byte, error) {
+	return nil, nil
+}
