+++ loggers/serial_writer_test.go
@@ -1,3 +1,6 @@
+//go:build !tinygo
+// +build !tinygo
+
 // Copyright 2022 Juan Pablo Tosso
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
